feat: Add scripts to generate converge, verify, and docs files
This commit adds three new scripts: gen-converge.sh, gen-verify.sh, and gen-docs.sh. These scripts generate YAML files for molecule converge, molecule verify, and documentation, respectively. The scripts iterate through the YAML files in the "defaults" directory and generate the necessary tasks and configurations based on the file names and contents.

The gen-converge.sh script generates the "converge.yml" file in the "molecule/default" directory. It creates tasks to install the corresponding roles using the "sgaunet.gh_role_installer" role and adds a sleep task to avoid errors.

The gen-verify.sh script generates the "verify.yml" file in the "molecule/default" directory. It includes default variables, checks if the "gh_role_installer" binary is present, and asserts its existence.

The gen-docs.sh script generates the "available_tools.md" file in the "docs" directory. It creates a markdown file listing the available tools that can be installed with self-contained variables. Each tool is linked to its corresponding GitHub repository, and the installation tasks are provided in a code block.

These scripts automate the generation of necessary files for molecule testing and documentation, improving the development workflow.